+++ libs/language-server/src/lib/validation/checks/column-id.spec.ts
@@ -19,7 +19,7 @@ import {
   extractColumnIdFromBlockProperty,
   parseHelper,
   validationAcceptorMockImpl,
-} from '../../utils';
+} from '../../../test';
 
 describe('column-id validation tests', () => {
   let parse: (
