+++ libs/language-server/src/lib/validation/checks/pipeline-definition.spec.ts
@@ -19,7 +19,7 @@ import {
   extractPipeline,
   parseHelper,
   validationAcceptorMockImpl,
-} from '../../utils';
+} from '../../../test';
 
 describe('pipeline-definition validation tests', () => {
   let parse: (
