+++ libs/language-server/src/lib/validation/checks/property-body.spec.ts
@@ -19,7 +19,7 @@ import {
   extractPropertyBodyFromBlock,
   parseHelper,
   validationAcceptorMockImpl,
-} from '../../utils';
+} from '../../../test';
 
 describe('property-body validation tests', () => {
   let parse: (
