+++ libs/language-server/src/test/index.ts
@@ -1,0 +1,6 @@
+// SPDX-FileCopyrightText: 2023 Friedrich-Alexander-Universitat Erlangen-Nurnberg
+//
+// SPDX-License-Identifier: AGPL-3.0-only
+
+export * from './langium-utils';
+export * from './utils';
